requirements.txt
# requirements.txt
# pip install -r requirements.txt

paho-mqtt==1.6.1
PyYAML==6.0.1
python-dotenv==1.0.1
Flask==3.0.3
Jinja2==3.1.4
boto3==1.34.84
botocore==1.34.84
gunicorn==22.0.0
requests==2.31.0
